mostly cleanup
Former-commit-id: d36d6a692e18a439e1ee573f482f3f3158130950

+++ common-tools/cnuphys/ced/src/main/java/cnuphys/ced/cedview/central/CentralZView.java
@@ -231,7 +231,7 @@ public class CentralZView extends CedView implements ChangeListener {
 
 		for (int layer = 1; layer <= 6; layer++) {
 			double r = BMTSectorItem.innerRadius[layer-1];
-			int region = (int) (layer+1)/2 - 1;
+			int region = (layer+1)/2 - 1;
 			double zmin;
 			double zmax;
 			Color color;
