+++ common-tools/cnuphys/magfield/src/main/java/cnuphys/magfield/RotatedCompositeField.java
@@ -71,7 +71,7 @@ public final class RotatedCompositeField extends CompositeField {
 //		System.out.println(" NEW: [ " + bx + ", " + by + ", " + bz + "] ");
 
 		result[0] = (float)(bx * _cos + bz * _sin);
-		result[1] = (float)(by);
+		result[1] = (by);
 		result[2] = (float)(bz * _cos - bx * _sin);
 //		System.out.println(" NEW: [ " + result[0] + ", " + result[1] + ", " +
 //		result[2] + "] ");
